SAFROLE fixes/refactor (#305)
* SAFROLE fixes/refactor

The primary goal here is to handle a new epoch change with SAFROLE. In
this case the old state alone is not sufficient, and we need to figure
out what the relevant components in the state will become in order to
properly produce blocks, submit tickets, and determine slot leaders.

- Refactor sealing and ticket functions to rather accept exactly what
  they need rather then the entire state, this makes it simpler to unit
test them and also allows us to push more complex logic to handle the
epoch change edge cases out to a few core node functions that will
eventually produce blocks, submit tickets and determine slot leaders.
- Refactor RingVerify/Prover to rather just operate on ValidatorsData
  directly since we can't just rely on the current state.
- Rename the TicketAccumulator to SealingKeys to avoid confusion.
- Update MaxTicketExtrinsicSize to 3 for tiny constants.
- Update the simulation test to properly handle the epoch change edge
  case.

* review feedback: add more comments

* Update internal/safrole/safrole.go

Co-authored-by: Leonidas Maroulis <pantrif@users.noreply.github.com>

---------

Co-authored-by: Leonidas Maroulis <pantrif@users.noreply.github.com>