ValidatorStatistics -> ActivityStatistics
Add initial ActivityStatistics structs which replace
ValidatorStatistics.

The primary purpose of this commit is to add enough of
ActivityStatistics to allow work on JSON encoding/decoding the state to
continue which requires these structs to exist.

The logic for computing ValidatorStatistics hasn't really changed. There
are no test vectors for the new core statistics and service statistics.
It seemed prudent to rather wait for vectors to be available before
starting work on the logic to compute them.

An attempt was made to update the test vectors to the latest despite the
above, but the changes to these test vectors affect other test vectors
too, and hence it makes more sense to update all the others with the
same minor changes.