Deserialize storage keys.
- Change the storage key used to a state key instead. This enables us to
  be able to deserialize the key later since serialization is lossy. We
don't need the original key since the PVM code that will make host calls
always knows the key it used. How we store that data and the key we use
to store it under is merely an implementation detail. As long as we can
set and get that data, it doesn't matter.
- Add storage key deserialization which is effectively just taking the
  serialized storage state key and it's value and adding it to the
appropriate services storage map. Serialization is now just as trivial.
- Update host calls to take the original key and produce the storage
  state key when writing/reading storage keys.