Fix PVM program blob parsing (#346)
The program blob now includes metadata first, and then the rest of the
program blob. Parse this first and update our validation checks.

Also fix the accumulate test. The accumulate code is running now so a
storage key is set by it. For now we need to ignore storage keys when
checking state equality until we update to the v0.6.5 vectors which do
contain storage keys as part of the vector's state.