Update state keys 0.6.7 (#381)
The last airity for the state key constructor function has changed. We
now do the key hashing inside this function, and the state serialisation
function merely passes the data to hash.

Because of this we can no longer tell the difference between storage
state keys and preimage meta keys anymore. The assumption now is that
the host shouldn't need to care how the keys are stored, as long as the
underlying host calls can read and write their data.

It's implied that all service dictionary keys should be stored in a
single map or KV DB. Each host call would know how to access it's key and
value.

So all functionality related to figuring out which type of service
dictionary key a given state key is has been removed. We can no longer
deserialize service dictionaries.

We will have to revisit this once we refactor the three service
dictionary maps into a single map of state key -> bytes.

Also
- The last airty of the state constructor function now takes plain
  bytes, so a special HashComponent type is no longer needed.