embed rust libraries (#418)
Embed our rust libraries for bandersnatch and erasurecoding so we can
produce a single Go binary.

The dynamic libraries are embedded using Go's native file embedding.
There is however, one catch. The libraries need to first be written back
to disk (in a temp directory) in order to be loaded by `purego`.

In the long run we'll probably want to switch to static libraries to
avoid this.