fix: safrole unexpected author
If we're in a fallback mode, and the author index is incorrect, we
should return an "unexpected author' error instead of just failing
signature validation.

This is because during fallback mode we know the author's public key,
and we know the public key that we expect in the sealing keys.

Also:
- Map epoch marker errors