fix: pvm fix for storage traces gas issue
In Go if you try shift a 64 bit value (for example) by more than 64 the result is zero.

This is different to the PVM spec which requires modding the amount you
shift by 64 first.

Doing an incorrect shift resulted in the program taking a slightly
different path and then causing us to consume more gas than needed.